meeting-agent/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-silero>=1.0.0
anthropic>=0.40.0
elevenlabs>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
